ACTIN-904: PR feedback
