+++ report/src/main/kotlin/com/hartwig/actin/report/interpretation/MolecularDriverEntry.kt
@@ -1,5 +1,6 @@
 package com.hartwig.actin.report.interpretation
 
+import com.hartwig.actin.Displayable
 import com.hartwig.actin.molecular.datamodel.DriverLikelihood
 import com.hartwig.actin.molecular.datamodel.evidence.ExternalTrial
 
@@ -12,4 +13,8 @@ data class MolecularDriverEntry(
     val externalTrials: Set<ExternalTrial> = emptySet(),
     val bestResponsiveEvidence: String? = null,
     val bestResistanceEvidence: String? = null
-)
+) : Displayable {
+    override fun display(): String {
+        return displayedName
+    }
+}
